updated user repo and controllers according to API doc

+++ src/main/java/com/depot/ims/repositories/SitesRepository.java
@@ -4,7 +4,9 @@ import com.depot.ims.models.Site;
 
 import java.util.List;
 
+import org.slf4j.helpers.CheckReturnValue;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
 public interface SitesRepository extends JpaRepository<Site, Integer>{
@@ -15,5 +17,9 @@ public interface SitesRepository extends JpaRepository<Site, Integer>{
     List<Site> findAll();
 
     @Query("SELECT s FROM Site s WHERE s.siteId=?1")
-    Site findBySiteId(Integer siteId);
+    Site findBySiteId(long siteId);
+
+//    @Modifying
+//    @Query("UPDATE Site s SET s.siteStatus = ?2 WHERE s.siteId=?1")
+//    Site updateStatus(long siteId, String status);
 }
